Speed up find performance on freeze/update
Instead of walking the directory tree for each repository type, simply
walk the tree one time mapping the correct function for each. This has
a pretty drastic speedup on my system even with an SSD over the other
method.

Also has the nice side effect that it reduces the code size of all the
freeze/update functions, not needing all the extra looping constructs.